Causa raiz: '/*' dentro de KDoc engolia arquivos inteiros (e o resto dos erros)
Toda referência a 'Unresolved reference BrowserDb / MatchPattern / listScripts / UserScript'
em 8 arquivos era cascata de UM problema: o Kotlin aninha comentário de bloco, e todo match
pattern escrito num KDoc contém '://*' — a barra com estrela abre um aninhado que é fechado
pelo '*/' de baixo, deixando o comentário externo pendurado até o fim do arquivo. O arquivo
para de ter declarações, o compilador não reclama dele, e reclama de quem usa.

No BrowserDb.kt isso comeu o UserScript inteiro (daí 'No parameter with name pattern/kind/
enabled/sourceExtension/runAtIdle found') e a classe BrowserDb. O extrator de anotações, depois
de aprender a tirar o carimbo de data do Actions, deixou aparecer o único erro honesto do lote:
'MatchPattern.kt:84:1 Syntax error: Unclosed comment'.

- KDocs reescritos: descrevem o curinga em vez de citá-lo, e explicam por quê.
- tools/kotlin-comment-check.py: tokenizador (string, raw string, char, // e /* aninhado) que
  aponta bloco pendurado; vira passo de CI antes de compilar, com ::error:: para o GitHub.
- ExtensionManager: o override de PromptDelegate#onInstallPromptRequest perde as projeções
  'out' nos três String[] — sobrepor método Java exige o tipo do parâmetro igual; com 'out' o
  155 também responde 'overrides nothing' (assinatura conferida no fonte da tag
  FIREFOX_155_0_1_RELEASE: GeckoResult<PermissionPromptResponse> onInstallPromptRequest(
  WebExtension, String[], String[], String[])).
- ManifestConverter.stableHash: 0x811C9DC5 não cabe em Int, o literal era inferido Long e o
  xor do meio do laço não compilava.

+++ app/src/main/kotlin/app/mobibrowser/core/ext/ManifestConverter.kt
@@ -316,7 +316,9 @@ object ManifestConverter {
 
     /** Hash curto determinístico (substituto do id da Chrome Web Store quando não há). */
     private fun stableHash(seed: String): String {
-        var h = 0x811C9DC5
+        // FNV-1a de 32 bits. O offset basis não cabe em Int: literal maior que
+        // Int.MAX_VALUE é inferido como Long e o resto do laço não compila.
+        var h = 0x811C9DC5.toInt()
         for (c in seed) {
             h = h xor c.code
             h *= 0x01000193
